Move two core-related api tests to core
