+++ detekt-core/src/test/resources/detekt.yml
@@ -1,0 +1,27 @@
+code-smell:
+  LongMethod:
+    active: true
+    threshold: 20
+  LongParameterList:
+    active: false
+    threshold: 5
+  LargeClass:
+    active: false
+    threshold: 70
+  InnerMap:
+    Inner1:
+      active: true
+    Inner2:
+      active: true
+
+style:
+  WildcardImport:
+    active: true
+  NoElseInWhenExpression:
+    active: true
+  MagicNumber:
+    active: true
+    ignoreNumbers: ['-1', '0', '1', '2']
+
+comments:
+  active: false
