+++ detekt-core/src/test/resources/pathFilters/multi-rule-with-excludes-on-ruleset.yml
