+++ detekt-core/src/test/resources/pathFilters/multi-rule-with-one-exclude-string.yml
